+++ docs/release-notes.md
@@ -8,6 +8,8 @@ nav_order: 8
 
 Major changes:
 
+- install: Add `--console` for configuring kernel and bootloader console
+- customize: Add `--dest-console` for configuring kernel and bootloader console
 - Support reading initrd images compressed with zstd
 - Add Fedora 38 signing key; drop Fedora 35 signing key
 
@@ -18,6 +20,7 @@ Minor changes:
 - Fix unlikely decompression error reading initrd
 - Add release notes to documentation
 - iso: Detect incomplete ISO files
+- Warn if console kargs could have used `--console`/`--dest-console` instead
 
 Internal changes:
 
@@ -28,7 +31,8 @@ Internal changes:
 Packaging changes:
 
 - Require Rust ≥ 1.58.0
-- Add dependency on `zstd` crate and `libzstd` shared library
+- Add dependencies on `textwrap` and `zstd` crates
+- Add dependency on `libzstd` shared library
 - Support `serde_yaml` 0.9
 - Remove non-Linux dependencies from vendor archive
 - Install example installer config file in `/usr/share/coreos-installer`